+++ templates/sync-function/template.js
@@ -86,15 +86,16 @@ function synctos(doc, oldDoc) {
   var theDocType = getDocumentType(doc, oldDoc);
 
   if (isValueNullOrUndefined(theDocType)) {
-    if (isDocumentMissingOrDeleted(oldDoc) && isDocumentMissingOrDeleted(doc)) {
-      // Attempting to delete a document that does not exist. This may occur when bucket access/sharing
+    if (doc._deleted) {
+      // Attempting to delete a document whose type is unknown. This may occur when bucket access/sharing
       // (https://developer.couchbase.com/documentation/mobile/current/guides/sync-gateway/shared-bucket-access.html)
-      // is enabled and the document was deleted via the Couchbase SDK. Skip everything else and simply assign the
-      // public channel
+      // is enabled and the document was deleted via the Couchbase SDK or if the document belongs to a type that existed
+      // in a previous version of the document definitions but has since been removed. Verify that the user has
+      // administrator access and then simply assign the public channel
       // (https://developer.couchbase.com/documentation/mobile/current/guides/sync-gateway/channels/index.html#special-channels)
-      // to the document so that users will get a 404 Not Found if they attempt to fetch (i.e. "view") the deleted
+      // to the document so that other users will get a 404 Not Found if they attempt to fetch (i.e. "view") the deleted
       // document rather than a 403 Forbidden.
-      requireAccess('!');
+      requireAccess([ ]); // This test can only be satisfied via the admin API
       channel('!');
 
       return;